nicer cli
