+++ examples/requirements.txt
@@ -13,3 +13,4 @@ streamlit
 elasticsearch
 pandas
 nlp
+fire
